A fix for the String class conversion of floats/doubles.
The issue is discussed here:
http://forum.arduino.cc/index.php?topic=349764.0

Apparently %f in sprintf() is not supported on this platform
(build option?). This is a work around using %d.

It doesn't properly support the width parameter. However,
the String methods (constructor and concat) both use a
fixed value of 4 for width.

Precision parameter is supported.