Add SODAQ boards.txt, platform.txt and variants
The files are a copy of v1.6.13-sodaq

The main difference with the previous sodaq branches is that we let all
the existing ArduinoCore content as is. The Arduino variants, the
bootloader is all pretty much the same. This way we are hoping to have less
work merging with the next ArduinoCore updates.

In this commit we have not yet added the bootloaders for the SODAQ boards.
This was to have a cleaner "source code" commit. The next commit will have
the bootloaders (bin and hex).